requirements.txt
opencv-python>=4.8.0,<5.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
pillow>=10.0.0,<11.0.0
twilio>=9.0.0,<10.0.0
python-dotenv>=1.0.0,<2.0.0
flask>=3.0.0,<4.0.0
flask-socketio>=5.3.0,<6.0.0
flask-cors>=4.0.0,<5.0.0
easyocr>=1.7.0,<2.0.0
pytesseract>=0.3.10,<0.4.0
tensorflow>=2.15.0;python_version<'3.12'
tf-keras>=2.15.0;python_version>='3.12'
eventlet>=0.33.0,<1.0.0
requests>=2.31.0,<3.0.0
